Used common Invoice components for generate & edit invoice
