+++ app/javascript/src/components/Invoices/Edit/index.tsx
@@ -9,13 +9,13 @@ import invoicesApi from "apis/invoices";
 import { sendGAPageView } from "utils/googleAnalytics";
 
 import Header from "./Header";
-import InvoiceTable from "./InvoiceTable";
-import InvoiceTotal from "./InvoiceTotal";
 
 import { unmapLineItems } from "../../../mapper/editInvoice.mapper";
+import CompanyInfo from "../common/CompanyInfo";
+import InvoiceDetails from "../common/InvoiceDetails";
+import InvoiceTable from "../common/InvoiceTable";
+import InvoiceTotal from "../common/InvoiceTotal";
 import { generateInvoiceLineItems } from "../common/utils";
-import CompanyInfo from "../CompanyInfo";
-import InvoiceDetails from "../Generate/InvoiceDetails";
 
 const EditInvoice = () => {
   const navigate = useNavigate();
@@ -112,13 +112,13 @@ const EditInvoice = () => {
           <InvoiceDetails
             currency={invoiceDetails.company.currency}
             clientList={invoiceDetails.companyClientList}
+            amount={amount}
             selectedClient={selectedClient || invoiceDetails.client}
             setSelectedClient={setSelectedClient}
-            amount={amount}
-            dueDate={dueDate || invoiceDetails.dueDate}
-            setDueDate={setDueDate}
             issueDate={issueDate || invoiceDetails.issueDate}
             setIssueDate={setIssueDate}
+            dueDate={dueDate || invoiceDetails.dueDate}
+            setDueDate={setDueDate}
             invoiceNumber={invoiceNumber}
             setInvoiceNumber={setInvoiceNumber}
             reference={reference}
@@ -140,17 +140,17 @@ const EditInvoice = () => {
           <InvoiceTotal
             currency={invoiceDetails.company.currency}
             newLineItems={selectedLineItems}
-            manualEntryArr={manualEntryArr}
-            setAmount={setAmount}
             amountPaid={amountPaid}
             amountDue={amountDue}
             setAmountDue={setAmountDue}
+            setAmount={setAmount}
             discount={discount}
             setDiscount={setDiscount}
             tax={tax || invoiceDetails.tax}
             setTax={setTax}
             showDiscountInput={!!invoiceDetails.discount}
             showTax={!!invoiceDetails.tax}
+            manualEntryArr={manualEntryArr}
           />
         </div>
       </React.Fragment>
