+++ app/javascript/src/components/Invoices/common/InvoiceDetails/index.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useRef } from "react";
 
 import dayjs from "dayjs";
+import { currencyFormat } from "helpers";
 import { PencilSimple } from "phosphor-react";
 
 import CustomDatePicker from "common/CustomDatePicker";
-import { currencyFormat } from "helpers/currency";
 
 import ClientSelection from "./ClientSelection";
